+++ src/index.html
@@ -23,7 +23,6 @@
       href="https://unpkg.com/swiper/swiper-bundle.min.css"
     />
     <link rel="stylesheet" href="./sass/index.scss" />
-    <script src="/src/js/page-scroll-btn.js" type="module"></script>
 
   </head>
   <body>
@@ -52,20 +51,6 @@
     <!-- Swiper JS -->
     <script src="https://unpkg.com/swiper/swiper-bundle.min.js" type="module"></script>
     <!-- Initialize Swiper -->
-    <script src="/src/js/product-mobile.js" type="module"></script>
-    <script src="/src/js/product2-mobile.js" type="module"></script>
-    <script src="/src/js/product3-mobile.js" type="module"></script>
-    <script src="/src/js/swiper.js" type="module"></script>
-    <!-- Counter -->
-
-
-    <script src="/src/js/numbers.js" type="module"></script>
-    <!-- About -->
-    <script src="./js/modal.js" type="module"></script>
-    <script src="./js/modal-form.js" type="module"></script>
-    <!-- Contacts -->
-    <script src="./js/contacts-modal.js" type="module"></script>
-     <script src="./js/location-modal.js" type="module"></script>
-
+    
   </body>
 </html>
